InputStreamReceiverBoltServer is organized
